Moves token related code into util so it can be reused.
